finish

+++ app/Http/Controllers/CommentController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Comment;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class CommentController extends Controller
 {
@@ -11,13 +12,24 @@ class CommentController extends Controller
     $comment=new Comment();
     $comment->content=request()->content;
     $comment->article_id=request()->article_id;
+    $comment->user_id=auth()->user()->id;
     $comment->save();
     return back();
 
     }
+
     public function delete($id){
-    $comment=Comment::find($id);
-    $comment->delete();
-    return back();
+        $comment=Comment::find($id);
+        if(Gate::denies('model-delete',$comment)){
+            return back()->with('error', 'Unauthorize');
+        }
+        $comment->delete();
+        return back();
     }
+
+
 }
+
+
+
+
